Type Definition pallet_offences::pallet::ReportsByKindIndex

pub type ReportsByKindIndex<T> = StorageMap<_GeneratedPrefixForStorageReportsByKindIndex<T>, Twox64Concat, Kind, Vec<u8>, ValueQuery>;

Enumerates all reports of a kind along with the time they happened.

All reports are sorted by the time of offence.

Note that the actual type of this mapping is Vec<u8>, this is because values of different types are not supported at the moment so we are doing the manual serialization.